+++ api/matlab/arrus/STASequence.m
@@ -1,33 +1,23 @@
 classdef STASequence < SimpleTxRxSequence
     % Synthetic Transmit Aperture Tx/Rx Sequence.
     %
-    % The length of the STA sequence is determined by the length of txApertureCenter.
-    % The txFocus and txAngle must be scalars or be the same length as txApertureCenter.
+    % :param rxCenterElement: vector of rx aperture center elements [element]. \
+    %   When empty, rx aperture center will be set to the first element.
+    % :param rxApertureCenter: vector of rx aperture center positions [m]. \
+    %   When empty, rx aperture center will be set to the first (0) element.
+    % :param rxApertureSize: size of the rx aperture [element]. \
+    %   When 0 and rxCenterElement and rxApertureCenter are empty, all  \
+    %   RX elements will be active.
     
     methods
         function obj = STASequence(varargin)
             obj = obj@SimpleTxRxSequence(varargin{:});
             
-            nTx = length(obj.txApertureCenter);
-            
-            % Manage the txFocus size
-            if length(obj.txFocus)~=nTx
-                if isscalar(obj.txFocus)
-                    obj.txFocus = obj.txFocus .* ones(1,nTx);
-                else
-                    error("ARRUS:IllegalArgument", ...
-                          "STASequence: txFocus must be scalar or be the same size as txApertureCenter");
-                end
-            end
-            
-            % Manage the txAngle size
-            if length(obj.txAngle)~=nTx
-                if isscalar(obj.txAngle)
-                    obj.txAngle = obj.txAngle .* ones(1,nTx);
-                else
-                    error("ARRUS:IllegalArgument", ...
-                          "STASequence: txAngle must be scalar or be the same size as txApertureCenter");
-                end
+            % Default rx aperture
+            if isempty(obj.rxCenterElement) && isempty(obj.rxApertureCenter) && obj.rxApertureSize == 0
+                obj.rxCenterElement = [];
+                obj.rxApertureCenter = 0.0 .* ones(1, length(obj.txAngle));
+                obj.rxApertureSize = "nElements";
             end
             
         end