+++ api/matlab/arrus/SimpleTxRxSequence.m
@@ -1,26 +1,33 @@
 classdef SimpleTxRxSequence < Operation
     % A sequence of Tx/Rx operations to perform on a device.
     %
+    % :param txCenterElement: vector of tx aperture center elements [element]
     % :param txApertureCenter: vector of tx aperture center positions [m]
     % :param txApertureSize: size of the tx aperture [element]
+    % :param rxCenterElement: vector of rx aperture center elements [element]
+    % :param rxApertureCenter: vector of rx aperture center positions [m].
+    % :param rxApertureSize: size of the rx aperture [element]
     % :param txFocus: tx focal length [m]
     % :param txAngle: tx angle [rad]
     % :param speedOfSound: speed of sound for [m/s]
     % :param txFrequency: tx frequency [Hz]
     % :param txNPeriods: number of sine periods in the tx burst (can be 0.5, 1, 1.5, etc.)
     % :param rxDepthRange: defines the end (if scalar) or
-    %                      the begining and the end (if two-element vector)
-    %                      of the acquisition expressed by depth range [m]
+    %   the begining and the end (if two-element vector) \ 
+    %   of the acquisition expressed by depth range [m]
     % :param rxNSamples: number of samples (if scalar) or 
-    %                      starting and ending sample numbers (if 2-element vector) 
-    %                      of recorded signal [sample]
-    % :param nRepetitions: number of repetitions of the sequence (positive integer)
+    %   starting and ending sample numbers (if 2-element vector) \ 
+    %   of recorded signal [sample]
+    % :param nRepetitions: number of repetitions of the sequence (positive \
+    %   integer). Can be a string "max" -- in this case, the maximum number \
+    %   of repetitions that can be performed on the system (taking into \ 
+    %   account the size of RAM, etc.) will be used.
     % :param txPri: tx pulse repetition interval [s]
     % :param tgcStart: TGC starting gain [dB]
     % :param tgcSlope: TGC gain slope [dB/m]
-    % :param fsDivider: sampling frequency divider. Should be positive int. 
-    %        Default value is equal to 1, which means sampling with 
-    %        the highest possible frequency (no decimation).
+    % :param fsDivider: sampling frequency divider. Should be positive int. \ 
+    %   Default value is equal to 1, which means sampling with \
+    %   the highest possible frequency (no decimation).
     % 
     % TGC gain = tgcStart + tgcSlope * propagation distance
     % TGC gain is limited to 14-54 dB, any values out of that range
@@ -29,7 +36,10 @@ classdef SimpleTxRxSequence < Operation
     properties
         txCenterElement (1,:) {mustBeFinite, mustBeReal}
         txApertureCenter (1,:) {mustBeFinite, mustBeReal}
-        txApertureSize (1,1) {mustBeFinite, mustBeInteger, mustBeNonnegative}
+        txApertureSize (1,1)
+        rxCenterElement (1,:) {mustBeFinite, mustBeReal}
+        rxApertureCenter (1,:) {mustBeFinite, mustBeReal}
+        rxApertureSize (1,1)
         txFocus (1,:) {mustBeNonNan, mustBeReal}
         txAngle (1,:) {mustBeFinite, mustBeReal}
         speedOfSound (1,1) {mustBeProperNumber}
@@ -37,7 +47,7 @@ classdef SimpleTxRxSequence < Operation
         txNPeriods (1,1) {mustBeInteger, mustBeProperNumber}
         rxDepthRange (1,:) {mustBeProperNumber}
         rxNSamples (1,:) {mustBeFinite, mustBeInteger, mustBePositive}
-        nRepetitions (1,1) {mustBeInteger, mustBePositive} = 1
+        nRepetitions (1,:) = 1
         txPri (1,1) double {mustBePositive} = 100e-6
         tgcStart (1,1)
         tgcSlope (1,1)
@@ -55,9 +65,38 @@ classdef SimpleTxRxSequence < Operation
                 obj.(varargin{i}) = varargin{i+1};
             end
             
-            mustBeXor(obj.rxDepthRange,obj.rxNSamples);
-            obj.rxDepthRange = mustBeLimit(obj.rxDepthRange,0);
-            obj.rxNSamples = mustBeLimit(obj.rxNSamples,1);
+            if ischar(obj.nRepetitions)
+                obj.nRepetitions = convertCharsToStrings(obj.nRepetitions);
+            end
+            
+            % Validate.
+            mustBeXor(obj,{'txCenterElement','txApertureCenter'});
+            if ~isempty(obj.rxCenterElement) || ~isempty(obj.rxApertureCenter)
+                mustBeXor(obj,{'rxCenterElement','rxApertureCenter'});
+            end
+            mustBeXor(obj,{'rxDepthRange','rxNSamples'});
+            obj.rxDepthRange = mustBeLimit(obj,'rxDepthRange',0);
+            obj.rxNSamples = mustBeLimit(obj,'rxNSamples',1);
+            
+            % Specific validations
+            mustBeIntOrStr(obj,'nRepetitions',1,"max");
+            mustBeIntOrStr(obj,'txApertureSize',0,"nElements");
+            mustBeIntOrStr(obj,'rxApertureSize',0,["nChannels","nElements"]);
+            
+            %% Check size compatibility of aperture/focus/angle parameters
+            nTx = max([	length(obj.txCenterElement) ...
+                        length(obj.txApertureCenter) ...
+                        length(obj.rxCenterElement) ...
+                        length(obj.rxApertureCenter) ...
+                        length(obj.txFocus) ...
+                        length(obj.txAngle) ]);
+            
+            obj.txCenterElement     = mustBeProperLength(obj.txCenterElement,nTx);
+            obj.txApertureCenter	= mustBeProperLength(obj.txApertureCenter,nTx);
+            obj.rxCenterElement     = mustBeProperLength(obj.rxCenterElement,nTx);
+            obj.rxApertureCenter	= mustBeProperLength(obj.rxApertureCenter,nTx);
+            obj.txFocus             = mustBeProperLength(obj.txFocus,nTx);
+            obj.txAngle             = mustBeProperLength(obj.txAngle,nTx);
             
         end
     end
@@ -70,39 +109,67 @@ function mustBeProperNumber(a)
     mustBeReal(a)
 end
 
-function mustBeXor(varargin)
+function mustBeXor(obj,fieldNames)
     
-    % input argument names
-    argNames = strings(1,nargin);
-    for iArg=1:nargin
-        argNames(iArg) = inputname(iArg);
-        if contains(argNames(iArg),'.')
-            argNames(iArg) = extractAfter(argNames(iArg),'.');
-        end
+    nField = length(fieldNames);
+    
+    fieldIsNonEmpty = false(1,nField);
+    for iField=1:nField
+        fieldIsNonEmpty(iField) = ~isempty(obj.(fieldNames{iField}));
     end
     
-    % error if more/less than one argument is non-empty
-    argIsNonEmpty = ~cellfun(@isempty,varargin);
-    if sum(argIsNonEmpty) ~= 1
+    if sum(fieldIsNonEmpty) ~= 1
         error("Arrus:params", ...
-            ['One and only one of: ' join(argNames,', ') ' must be defined.'])
+            ['One and only one of {' char(join(fieldNames,', ')) '} must be defined.'])
     end
 end
 
-function argOut = mustBeLimit(argIn,defLo)
+function mustBeIntOrStr(obj,fieldName,intDomain,strDomain)
+    % intDomain: 0-nonnegative integers; 1-positive integers
+    % strDomain: array of accepted string values
     
-    argName = inputname(1);
-    if contains(argName,'.')
-        argName = extractAfter(argName,'.');
+    argIn = obj.(fieldName);
+    
+    if isnumeric(argIn)
+        mustBeInteger(argIn);
+        if intDomain
+            mustBePositive(argIn);
+        else
+            mustBeNonnegative(argIn);
+        end
+    elseif isstring(argIn)
+        mustBeMember(argIn,strDomain);
+    else
+        error(['Unhandled data type of ' fieldName ': ', ...
+            class(argIn)])
     end
+    
+end
+
+function argOut = mustBeProperLength(argIn,propLength)
+    
+    if isempty(argIn) || length(argIn)==propLength
+        argOut = argIn;
+    elseif isscalar(argIn)
+        argOut = argIn .* ones(1,propLength);
+    else
+        error("ARRUS:IllegalArgument", ...
+                          "Incompatible parameter length");
+    end
+    
+end
+
+function argOut = mustBeLimit(obj,fieldName,defLo)
+    
+    argIn = obj.(fieldName);
     
     if isempty(argIn)
         argOut = argIn;
     else
         % check the size
-        if length(argIn(:))>2
+        if numel(argIn)>2
             error("Arrus:params", ...
-                ['Value assigned to ' argName ' property ', ...
+                ['Value assigned to ' fieldName ' property ', ...
                 'should be a scalar or two-element vector'])
         end
         
@@ -116,7 +183,7 @@ function argOut = mustBeLimit(argIn,defLo)
         % check the ascending order
         if argOut(2) <= argOut(1)
             error("Arrus:params", ...
-                ['The ' argName ' property should have ascending order']);
+                ['The ' fieldName ' property should have ascending order']);
         end
     end