+++ docs/content/misc/release_notes.rst
@@ -1,6 +1,17 @@
 Release notes
 =============
 
+Version 0.4.2
+-------------
+
+- MATLAB API:
+
+    - ``rxApertureCenter``, ``rxCenterElement``, and ``rxApertureSize`` parameters added \
+      to control the rx aperture. Use them the same way as parameters for tx aperture control. \
+      Now the rx aperture is fully defined by those parameters and sequence type no longer affects it.
+    - rx delay is set to zero to allow for shallow region imaging.
+    - ``nRepetitions``: set it to "max" to acquire maximum number of repetitions allowable.
+
 Version 0.4.1
 -------------